chore: Enable Cyclop linter
Quite some refactoring here. I haven't been able to cover everything (there are a few nolint annotations).

Be careful in reviewing this PR as some refactors are of methods that reuse the variable multiple times (e.g. reassigning to err). I tried to keep the functionality as is as much as possible.

Signed-off-by: Lennard Eijsackers <lennardeijsackers92@gmail.com>