chore: Refactor host in scope and seperate to new package topology
*Motivation:*
Right now a lot of the code base in chproxy resides in a shared main package. This has lead to a lot of coupled code and a code base that is very hard to read.

This PR is a continuation of the work started with the heartbeat to move code away from the main package, decouple the code and improve readability.

*Additions and Changes:*
- Create a new `Node` struct in the new `topology` package that exposes methods from the previous `host` struct. However, it doesn't expose internal state.
- Improve the Node code by using more modern constructs such as `atomic.Bool`.
- Update the scope package and every usage of `host` with `topology.Node`.
- Include a new test case for `Node.StartHeartbeat`

*Notes:*
Due to the coupled nature of the code around scope, I didn't see an opportunity to do this incrementally. The PR will sadly be large and hard to review.

Signed-off-by: Lennard Eijsackers <lennardeijsackers92@gmail.com>